+++ src/custom/CFConst.gd
@@ -116,6 +116,7 @@ const PATH_TOKENS := PATH_ASSETS + "tokens/"
 # The scripting engine functionality with your own tasks,
 # Point this to your own script file.
 const PATH_SCRIPTING_ENGINE := PATH_CORE + "ScriptingEngine.gd"
+const PATH_PER_ENGINE := PATH_CORE + "ScriptPer.gd"
 # The amount of distance neighboring cards are pushed during card focus
 #
 # It's based on the card width. Bigger percentage means larger push.